Move the field form into separate Component.
