+++ Ryhma57/src/main/java/ryhma57/gui/Window.java
@@ -16,9 +16,6 @@ public class Window extends JFrame implements ActionListener {
     
     private Application app;
     private JLabel infoLabel;
-
-    /* this is bit ugly */
-    private EnumMap<BibtexReferenceField, JFormattedTextField> fields;
     
     private String[] referenceTypes = {"Book", "Inproceedings", "Article"};
 
@@ -26,56 +23,17 @@ public class Window extends JFrame implements ActionListener {
         this.app = app;
         this.setTitle("Application");
         /* set some default size so that window isn't 0x0 size */
- /*XXX we could load window geometry from config file */
+        /*XXX we could load window geometry from config file */
         this.setSize(600, 480);
         this.setDefaultCloseOperation(EXIT_ON_CLOSE);
 
-        this.fields = new EnumMap<>(BibtexReferenceField.class);
-        this.createDialog();
+        this.createComponents();
     }
 
-    private void generateField(JPanel labelPane, JPanel inputPane, BibtexReferenceField field) {
-        JFormattedTextField input;
-
-        labelPane.add(new JLabel(field.getName() + ":"));
-        input = new JFormattedTextField();
-        input.setColumns(10);
-        this.fields.put(field, input);
-        inputPane.add(input);
-    }
-
-    private void createDialog() {
-        JFormattedTextField input;
-        JPanel mainPane = new JPanel(new BorderLayout());
-        JPanel labelPane = new JPanel(new GridLayout(0, 1));
-        JPanel inputPane = new JPanel(new GridLayout(0, 1));
-
-        mainPane.setBorder(BorderFactory.createTitledBorder("Create a reference"));
-
-        generateField(labelPane, inputPane, BibtexReferenceField.ID);
-        generateField(labelPane, inputPane, BibtexReferenceField.AUTHOR);
-        generateField(labelPane, inputPane, BibtexReferenceField.EDITOR);
-        generateField(labelPane, inputPane, BibtexReferenceField.TITLE);
-        generateField(labelPane, inputPane, BibtexReferenceField.PUBLISHER);
-        generateField(labelPane, inputPane, BibtexReferenceField.YEAR);
-
-        generateField(labelPane, inputPane, BibtexReferenceField.VOLUME);
-        generateField(labelPane, inputPane, BibtexReferenceField.NUMBER);
-        generateField(labelPane, inputPane, BibtexReferenceField.SERIES);
-        generateField(labelPane, inputPane, BibtexReferenceField.ADDRESS);
-        generateField(labelPane, inputPane, BibtexReferenceField.EDITION);
-        generateField(labelPane, inputPane, BibtexReferenceField.MONTH);
-        generateField(labelPane, inputPane, BibtexReferenceField.NOTE);
-
-        mainPane.add(labelPane, BorderLayout.CENTER);
-        mainPane.add(inputPane, BorderLayout.LINE_END);
-        this.add(mainPane, BorderLayout.NORTH);
-
+    private void createComponents() {
         JButton button;
-        button = new JButton("Save");
-        button.addActionListener(this);
-        button.setActionCommand(CREATE);
-        mainPane.add(button, BorderLayout.SOUTH);
+        JPanel mainPane = new FieldsForm(this.app);
+        this.add(mainPane, BorderLayout.NORTH);
 
         button = new JButton("Generate BibTeX file");
         button.addActionListener(this);
@@ -84,7 +42,7 @@ public class Window extends JFrame implements ActionListener {
         this.infoLabel = new JLabel("");
         this.add(this.infoLabel, BorderLayout.CENTER);
 
-        this.add(button, BorderLayout.SOUTH);        
+        this.add(button, BorderLayout.SOUTH);
         
         JComboBox referenceList = new JComboBox(referenceTypes);
         referenceList.addActionListener(this);
@@ -92,31 +50,18 @@ public class Window extends JFrame implements ActionListener {
         mainPane.add(referenceList, BorderLayout.NORTH);
     }
 
+    public void setErrorMessage(String error) {
+        Timer timer;
+
+        this.infoLabel.setText("The reference was saved");
+        timer = new Timer(10000, this);
+        timer.start();
+        timer.setActionCommand(CLEAR_LABEL);
+    }
+
     @Override
     public void actionPerformed(ActionEvent event) {
-        if (event.getActionCommand().equals(CREATE)) {
-            EnumMap<BibtexReferenceField, String> set;
-            String error;
-            Timer timer;
-            set = new EnumMap<>(BibtexReferenceField.class);
-
-            for (BibtexReferenceField field : this.fields.keySet()) {
-                JFormattedTextField input;
-                input = this.fields.get(field);
-                set.put(field, input.getText());
-            }
-            error = this.app.createNewBookReference(set);
-            if (error != null) {
-                this.infoLabel.setText(error);
-                return;
-            }
-
-            this.infoLabel.setText("The reference was saved");
-            timer = new Timer(10000, this);
-            timer.start();
-            timer.setActionCommand(CLEAR_LABEL);
-
-        } else if (event.getActionCommand().equals(GENERATE)) {
+        if (event.getActionCommand().equals(GENERATE)) {
             this.app.generateBibTex();
             JOptionPane.showMessageDialog(this,
                     "BibTex file was generated.");
